refactor(evals): extract shared helpers to evals-utils to remove 4 clones
Move three helpers into evals-utils.ts that were duplicated between
run.ts and eval-compare.ts:
- buildToolDefsSubset(): build tool definition subset for cache key
- assertAndRecord(): run assertions and construct RunResult
- makeFailedRunResult(): build a failed RunResult from an error catch

Both files now import from evals-utils instead of repeating the logic.
Also fixes log lines in eval-compare to use result.passed / result.assertionFailed.
